Add new skew strategy for estimating chunking and make default
The original strategy produced a chunk that was balanced according to
the expected data dimensions - this is reasonable if processors are
likely to slice over different dimensions but it is not so good for
writing or processing single detector images at a time.

The new strategy ("skew") weights the chunk dimensions toward the last
dimensions, reducing earlier dimensions to one before moving to the next
(or until the chunk is under the target size - usually 1MB).

This is made the default as it is more likely to cater for the common
case when people forget to specify chunking.

Signed-off-by: Charles Mita <charles.mita@diamond.ac.uk>