Incrementally test link path in NexusFileHDF5 when validating.
H5Lexists "fails" if any part of the path to test does not exist (except
the last name). Whilst we can catch the resulting exception and just
return false, this can result in a noisy message being printed to stderr
by the library, which we would like to avoid.

We follow the recommendation in the HDF5 reference to test the path
incrementally.

Signed-off-by: Charles Mita <charles.mita@diamond.ac.uk>