Regenerate Nexus file and fix compilation issues (#570)
* [DAQ-442] regenerated nexus base classes from NXDL definitions

fixed issue with clashing group/field names in NXsample for sample_component

Signed-off-by: Matthew Dickie <matthew.dickie@diamond.ac.uk>

* [DAQ-442] fix call to NXprocess.getAllCollection() in NXcanSASValidator

this method doesn't exist in NXprocess as NXcollection is not listed
as a possible child group. Where this is the case the call should be
to group.getChildren(NXcollection.class). This is fixed for the general
case in the XSLT generation of the validator classes.

Signed-off-by: Matthew Dickie <matthew.dickie@diamond.ac.uk>