[DAQ-822] NexusFileHDF5.populateGroupNode() check that external file can be opened
If we are trying to link a Nexus file to an external file (in SWMR
mode) which itself has external links, the open can fail because it
seems that the h5 library does not use SWMR to open the "child" files.

This fix puts in an extra check to see whether the external file can
be opened and, if it cannot, it continues as if the file does not exist.

Signed-off-by: Anthony Hull <anthony.hull@diamond.ac.uk>